requirements.txt
Flask
Flask-Cors
SQLAlchemy
Werkzeug
python-dotenv
pandas
reportlab
PyMySQL
psycopg2-binary
pyodbc
gunicorn